サインアップ、ログインのロジックを外部ファイル化
